refactor: Extract useProjectFilters from Projects.tsx
