+++ src/components/content/Projects/Projects.tsx
@@ -1,195 +1,25 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
 import { useNotionSectionData } from "@/hooks/useNotionSectionData";
+import { useProjectFilters } from "@/hooks/useProjectFilters";
 import type { NotionData } from "@/types/content";
-import { generateTagColors } from "@/utils/colorUtils";
 import { cn } from "@/utils/commonUtils";
-import {
-  createProjectEffect,
-  DEFAULT_PROJECT_EFFECT,
-  type MoireEffectPreset,
-} from "@/utils/moireEffectPresets";
-import { reconcileProjectFilters } from "@/utils/reconcileProjectFilters";
-import PixelCanvas from "../../effects/PixelCanvas/PixelCanvas";
+import { createProjectEffect } from "@/utils/moireEffectPresets";
 import { NotionSectionSkeleton } from "../shared/NotionSectionSkeleton";
+import ProjectCard from "./ProjectCard";
 
-interface ProjectCardProps {
-  title: string;
-  hook: string;
-  detail: string;
-  slug: string;
-  link?: string | null;
-  keywords: string[];
-  date: string | number | null;
-  image?: string | null;
-  tagColors?: Record<string, string>;
-  primaryTagColor?: string;
-  className?: string;
-  effect?: MoireEffectPreset;
-}
-
-function ProjectCard({
-  title,
-  hook,
-  detail,
-  slug,
-  link,
-  keywords,
-  date,
-  image,
-  tagColors,
-  primaryTagColor,
-  className = "",
-  effect = DEFAULT_PROJECT_EFFECT,
-}: ProjectCardProps) {
-  const [isClicked, setIsClicked] = useState(false);
-
-  const handleClick = (e: React.MouseEvent) => {
-    if (!isClicked) {
-      e.preventDefault();
-      setIsClicked(true);
-    }
-  };
-
-  const _link = link ? (
-    <div className="projects__card__label projects__card__link">Link</div>
-  ) : null;
-
-  return (
-    <a
-      href={link || undefined}
-      target={link ? "_blank" : undefined}
-      rel={link ? "noreferrer" : undefined}
-      className={cn(
-        `projects__card ${className}`.trim(),
-        image && "has-image",
-        isClicked && "is-expanded",
-      )}
-      key={slug}
-      onClick={handleClick}
-    >
-      <PixelCanvas
-        className="projects__card__pixel-canvas"
-        colors={effect.colors}
-        gap={effect.gap}
-        speed={effect.speed}
-      />
-      <div className="projects__card__content">
-        <div className="projects__card__meta">
-          <p className="projects__card__year">{date ?? ""}</p>
-          <div className="projects__card__keywords">
-            {_link}
-            {keywords.map((keyword) => (
-              <div
-                key={keyword}
-                className="projects__card__label"
-                style={{
-                  backgroundColor:
-                    tagColors?.[keyword] ||
-                    primaryTagColor ||
-                    "rgba(255, 255, 255, 0.25)",
-                }}
-              >
-                {keyword}
-              </div>
-            ))}
-          </div>
-        </div>
-        <h3>{title}</h3>
-        <p className="projects__card__hook">{hook}</p>
-        <div
-          className={cn(
-            "projects__card__detail-wrapper",
-            isClicked && "show-text",
-          )}
-        >
-          <div className="projects__card__detail-inner">
-            <p className="projects__card__detail">{detail}</p>
-          </div>
-        </div>
-        {image && (
-          <img
-            src={image}
-            className="project-image"
-            alt="Project"
-            width={16}
-            height={9}
-          />
-        )}
-      </div>
-    </a>
-  );
-}
 interface ProjectsProps {
   db?: Pick<NotionData, "projects">;
 }
 
 function Projects({ db: propsDb }: ProjectsProps = {}) {
-  const [activeFilters, setActiveFilters] = useState<string[]>([]);
-  const [tagColors, setTagColors] = useState<Record<string, string>>({});
   const { db, isLoading } = useNotionSectionData(propsDb);
 
   const projectsData = useMemo(
     () => (Array.isArray(db?.projects) ? db.projects : []),
     [db?.projects],
   );
-  const allKeywords = useMemo(
-    () =>
-      Array.from(
-        new Set(
-          projectsData.flatMap((project) =>
-            Array.isArray(project.keywords) ? project.keywords : [],
-          ),
-        ),
-      ).filter(
-        (keyword): keyword is string =>
-          typeof keyword === "string" && keyword.trim().length > 0,
-      ),
-    [projectsData],
-  );
-
-  const syncTagColors = useCallback(() => {
-    setTagColors(generateTagColors(allKeywords));
-  }, [allKeywords]);
-
-  const syncFilters = useCallback(() => {
-    setActiveFilters((prevFilters) =>
-      reconcileProjectFilters(prevFilters, allKeywords),
-    );
-  }, [allKeywords]);
-
-  useEffect(() => {
-    syncTagColors();
-    syncFilters();
-  }, [syncTagColors, syncFilters]);
-
-  useEffect(() => {
-    document.body.addEventListener("theme-changed", syncTagColors);
-
-    return () => {
-      document.body.removeEventListener("theme-changed", syncTagColors);
-    };
-  }, [syncTagColors]);
-
-  const toggleFilter = useCallback(
-    (filter: string) => {
-      setActiveFilters((prevFilters) => {
-        if (prevFilters.includes(filter)) {
-          if (prevFilters.length === 1) {
-            return [...allKeywords];
-          }
-          return prevFilters.filter((f) => f !== filter);
-        }
-
-        return [...prevFilters, filter];
-      });
-    },
-    [allKeywords],
-  );
-
-  const activeFiltersSet = useMemo(
-    () => new Set(activeFilters),
-    [activeFilters],
-  );
+  const { allKeywords, tagColors, activeFiltersSet, toggleFilter } =
+    useProjectFilters(projectsData);
 
   const project_cards = projectsData.map((projectProps, index) => {
     const primaryKeyword = projectProps.keywords[0] || "";
